+++ internal/router/interchain_test.go
@@ -107,7 +107,7 @@ func mockBlock(blockNumber uint64, txs []*pb.Transaction) *pb.Block {
 func mockTx(data *pb.TransactionData) *pb.Transaction {
 	return &pb.Transaction{
 		Data:  data,
-		Nonce: rand.Int63(),
+		Nonce: uint64(rand.Int63()),
 	}
 }
